gha: Enable KBS images for s390x
This commit introduces a job matrix to allow a s390x self-hosted runner
to run alongside the existing x86_64 one.

Additionally, two new steps (e.g., {pre,post} action) are introduced to
manage the self-hosted runner since it is not provisioned instantly for
CI.

To run the build steps `kbs` and `kbs-grpc-as` in parallel, the build step
is generalized into one step and the relevant configuration values are
parameterized into the job matrix.

Signed-off-by: Hyounggyu Choi <Hyounggyu.Choi@ibm.com>